+++ examples/next-openai/app/api/use-chat-reasoning/route.ts
@@ -1,5 +1,5 @@
-import { deepseek } from '@ai-sdk/deepseek';
-import { streamText } from 'ai';
+import { fireworks } from '@ai-sdk/fireworks';
+import { extractReasoningMiddleware, streamText, wrapLanguageModel } from 'ai';
 
 // Allow streaming responses up to 30 seconds
 export const maxDuration = 30;
@@ -7,10 +7,30 @@ export const maxDuration = 30;
 export async function POST(req: Request) {
   const { messages } = await req.json();
 
+  console.log(JSON.stringify(messages, null, 2));
+
   const result = streamText({
-    model: deepseek('deepseek-reasoner'),
+    model: wrapLanguageModel({
+      model: fireworks('accounts/fireworks/models/deepseek-r1'),
+      middleware: extractReasoningMiddleware({ tagName: 'think' }),
+    }),
     messages,
   });
+
+  // const result = streamText({
+  //   model: deepseek('deepseek-reasoner'),
+  //   messages,
+  // });
+
+  // const result = streamText({
+  //   model: anthropic('research-claude-flannel'),
+  //   messages,
+  //   providerOptions: {
+  //     anthropic: {
+  //       thinking: { type: 'enabled', budgetTokens: 12000 },
+  //     },
+  //   },
+  // });
 
   return result.toDataStreamResponse({
     sendReasoning: true,