+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/svelte
 
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [ddf9740]
+  - @ai-sdk/ui-utils@1.1.16
+  - @ai-sdk/provider-utils@2.1.10
+
 ## 1.1.18
 
 ### Patch Changes