+++ packages/togetherai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/togetherai
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [ddf9740]
+  - @ai-sdk/provider@1.0.9
+  - @ai-sdk/openai-compatible@0.1.12
+  - @ai-sdk/provider-utils@2.1.10
+
 ## 0.1.12
 
 ### Patch Changes